Move nav bar into ui folder
