+++ src/components/ui/NavigationBar.vue
